Fixed the bug where particles go through the bottom of the screen before bounding
Why: the particles could go through the bottom of the screen

Notes: the issue was that I was passing the frame rather than the JPanel, the frame includes the top bar with the title and that was messign up the sizing. I fixed it by changing all of the frame. to this.